added iplotNET function
Added a slightly modified version of the plotNET function from the
pmagplotlib.py library. This modification to the function does not
print out the version of pmagpy used to generate the figure as that
added additional white space to the bottom of the figure taking up more
room in the notebook. This version is also different in that it does
not print a box around the equal area plot.